Changed float to double to prevent compilation error by buildbot.

+++ tracking/datcon/fpgaDATCON/findlets/include/DATCONSVDClusterLoaderAndPreparer.h
@@ -67,13 +67,13 @@ namespace Belle2 {
     /// shift in r-phi to create windmill structure, in µm
     const std::array<int, 4> rPhiShiftsOfLayers = { -4680, -10780, -11036, -19076};
     /// phi-value of the first ladder in each layer, i.e. sensors X.1.Y, in rad
-    const std::array<float, 4> initialAngle = {(342.195 - 360.) / 180. * M_PI, 8. / 180. * M_PI, -8. / 180. * M_PI, -4. / 180. * M_PI};
+    const std::array<double, 4> initialAngle = {(342.195 - 360.) / 180. * M_PI, 8. / 180. * M_PI, -8. / 180. * M_PI, -4. / 180. * M_PI};
     /// angle difference between two consecutive ladders, in rad
-    const std::array<float, 4> angleStep = {2. * M_PI / 7., M_PI / 5., M_PI / 6., M_PI / 8.};    // in rad
+    const std::array<double, 4> angleStep = {2. * M_PI / 7., M_PI / 5., M_PI / 6., M_PI / 8.};    // in rad
     /// cosine values of the slanted sensors
-    const std::array<float, 4> cosSlantedAngles = {1, cos(0.207694180987), cos(0.279252680319), cos(0.368264472171)};
+    const std::array<double, 4> cosSlantedAngles = {1, cos(0.207694180987), cos(0.279252680319), cos(0.368264472171)};
     /// sine values of the slanted sensors
-    const std::array<float, 4> sinSlantedAngles = {1, sin(0.207694180987), sin(0.279252680319), sin(0.368264472171)};
+    const std::array<double, 4> sinSlantedAngles = {1, sin(0.207694180987), sin(0.279252680319), sin(0.368264472171)};
 
     /// shift along z of the L3 senosrs, in µn
     const std::array<int, 2> zShiftL3 = {92350, -32650};
@@ -84,10 +84,10 @@ namespace Belle2 {
     /// shift along z of the L6 senosrs, in µn
     const std::array<int, 5> zShiftL6 = {303471, 182060, 57060, -67940, -192940};
 
-    /// convert float to long int for more similarity to the FPGA implementation
+    /// convert double to long int for more similarity to the FPGA implementation
     /// @param value to be converted
     /// @param power multiply value by 10^power
-    inline long convertToInt(float value, int power)
+    inline long convertToInt(double value, int power)
     {
       long factor = (long)pow(10, power);
       return round(factor * value);
